% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPeakFilling.R
\name{batchPeakFilling}
\alias{batchPeakFilling}
\title{Function to do peak filling (xcms) with restored RT values}
\usage{
batchPeakFilling(xcmsObject, unStructureBatchesObject)
}
\arguments{
\item{xcmsObject}{XCMS object to be peak filled}

\item{unStructureBatchesObject}{XCMS object after retreived after the unStructureBatches fumction}
}
\value{
xcmsObject A peak filled xcms object
}
\description{
Function to do peak filling (xcms) with restored RT values
}
\examples{


}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
