% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blanco.check.R
\name{blanco.check}
\alias{blanco.check}
\title{Blanco filter}
\usage{
blanco.check(DataMatrix, subclasses.list = NULL, blanco.entry = NULL,
  blanco.data = NULL, Blanco.multiplier = 10,
  ignore.missing.values = TRUE, labels = NULL,
  feature.orientation = "columns", groups.ok.threshold = 1)
}
\arguments{
\item{DataMatrix}{A matrix with feature groups as columns and samples for rows.}

\item{subclasses.list}{A list with either A) in each list entry a number of integers or 
numerics corresponding to the rows that form the subclasses to be checked, or B) in each 
list entry characters or factors corresponding to those found in the 'labels' attribute.}

\item{blanco.entry}{The label of the blanco samples (as to be found in the 'labels' attribute) or 
a vector of row numbers corresponding to the rows with blanco samples.}

\item{blanco.data}{If the blanco data is not in 'DataMatrix' but in an additional matrix than it can
be provided with this entry.}

\item{Blanco.multiplier}{The maximal RSD value (sd/mean) allowed in a single subclass.}

\item{ignore.missing.values}{Whether to ignore the missing values. If set to FALSE 
the calculation of the median intensities will result in a lot 'NAs' or 'NaNs'}

\item{labels}{A vector of labels which can be used to group the samples together in subclasses. 
This attribute is necessary if 'subclasses.list' consists of character or factor entries.}

\item{feature.orientation}{Indicates whether the features can be found in the columns (default) or 
in the rows. With the default setting every row corresponds to a sample, and every column to a feature/group}

\item{groups.ok.threshold}{The minimal amount of subclasses (as defined in subclasses.list) that have to pass the test}
}
\value{
blanco.ok A vector with the indices of groups/features which passed the blanco filter.
}
\description{
function only lets groups of samples/subclasses pass for which at least one of the 
subclasses has a median value x times (10 default) larger than the maximal blanco value
}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
