% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAMERA.accumulator.R
\name{CAMERA.accumulator}
\alias{CAMERA.accumulator}
\title{CAMERA.accumulator}
\usage{
CAMERA.accumulator(XCMS.object, polarity = NULL, perfwhm = 0.6,
  ppm_threshold = 10, ppm_threshold.multiplier = 1.5, mzabs = 0.02,
  intval = "into", maxcharge = 2, nCPU = -1, Standards.list = NULL)
}
\arguments{
\item{XCMS.object}{xcmsSet object}

\item{polarity}{xcmsSet polarity}

\item{perfwhm}{xcms perfwhm parameter}

\item{ppm_threshold}{xcms ppm_threshold parameter}

\item{ppm_threshold.multiplier}{parameter with which the ppm_threshold is multiplied}

\item{mzabs}{xcms mzabs parameter}

\item{intval}{xcms intval parameter}

\item{maxcharge}{xcms maxcharge parameter}

\item{nCPU}{The number of CPUs to use. Default is maximum available minus 2.}

\item{Standards.list}{Yet to implement}
}
\description{
Function to accumulate CAMERA results for all combinations of given parameters.
}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
