% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EICsBatchPlotter.R
\name{EICsBatchPlotter}
\alias{EICsBatchPlotter}
\title{Get EIC's and plot in ggplot}
\usage{
EICsBatchPlotter(XCMSobject, groups = NULL, sample_names = NULL,
  sample_indices = NULL, phenotype_classname = NULL,
  plot_samples = FALSE, ...)
}
\arguments{
\item{XCMSobject}{A single XCMS object}

\item{groups}{(optional) The group indices for which to obtain an EIC. Default is all groups.}

\item{sample_names}{(optional) The names of samples (from XCMSobject@phenoData) to include in the EIC plots.}

\item{sample_indices}{(optional) The indices of samples (from XCMSobject@phenoData) to include in the EIC plots.}

\item{phenotype_classname}{Variable name (from XCMSobject@phenoData) containing class information for the plot colours.}

\item{plot_samples}{Whether to plot individual samples (default = FALSE: sample groups plotted in distinct colours)}

\item{...}{Other variables to be passed along to \link[xcms]{getEIC}}
}
\value{
EICresults A list with the data for plotting, and the ggplot objects.
}
\description{
Get EIC's and plot in ggplot
}
\author{
Matthias Cuyckx, \email{matthias.cuykx@uantwerpen.be}

Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
