% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureQualityScore.R
\name{FeatureQualityScore}
\alias{FeatureQualityScore}
\title{Feature Quality score}
\usage{
FeatureQualityScore(XCMSobject, indices.to.keep = NULL)
}
\arguments{
\item{XCMSobject}{An xcmsSet object}

\item{indices.to.keep}{Vector with indices to keep. Either numeric indices or a vector with TRUE/FALSE (same length as nr. of groups in XCMSobject)}
}
\value{
A vector with QC scores (between 0 and 1)
}
\description{
Calculate the feature quality score based on the presence of non failed gaussian fits.
}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
