% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBatchStructure.R
\name{getBatchStructure}
\alias{getBatchStructure}
\title{Get the Batch processing Structure. Necessary for merging batches again after analysis}
\usage{
getBatchStructure(xcmsSets.Post.batchSplit)
}
\arguments{
\item{xcmsSets.Post.batchSplit}{list of xcmsSet objects after Batch splitting (BatchSplitR function)}
}
\value{
BatchProcessingStructure A data frame with the list structure (quantifies the list nesting)
}
\description{
Get the Batch processing Structure. Necessary for merging batches again after analysis
}
\examples{


}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
