% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFoldChanges.R
\name{getFoldChanges}
\alias{getFoldChanges}
\title{calculate fold changes}
\usage{
getFoldChanges(DataMatrix, classlabels, baseClass = NULL,
  classOfInterest = NULL, datatype = "ints")
}
\arguments{
\item{DataMatrix}{The data matrix}

\item{classlabels}{the class labels (nr of vector elements must be equal to number of rows in DataMatrix)}

\item{baseClass}{The baseclass for the fold change clauculation}

\item{classOfInterest}{The class of interest for the fold change calculation}

\item{datatype}{Indication whether the Datamatrix is intensities (ints) or log2 values (log2s).}
}
\value{
foldChanges A vector with the fold changes
}
\description{
Filter out isotopes based on the PUTMEDID concept
}
\examples{


}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
