% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glog.R
\name{glog}
\alias{glog}
\title{generalized logarithm scaling}
\usage{
glog()
}
\value{

}
\description{
transform the data to glog.
}
\examples{

 
}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
