% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.filter.R
\name{group.filter}
\alias{group.filter}
\title{Group Filtering}
\usage{
group.filter(XCMSobject, max.RT.diff = NULL, max.ppm.diff = NULL,
  max.dupl.prop = NULL, max.missing.prop = NULL,
  thresholds.to.pass = NULL)
}
\arguments{
\item{XCMSobject}{An xcmsSet object after peak picking.}

\item{max.RT.diff}{The maximal allowed RT width for a group}

\item{max.ppm.diff}{The maximal allowed ppm difference within a group}

\item{max.dupl.prop}{The maximal allowed proportion of duplicated peaks}

\item{max.missing.prop}{The maximal allowed proportion of missing peaks}

\item{thresholds.to.pass}{The number of thresholds to pass (default is all)}
}
\value{

}
\description{
After grouping and peak filling, this function filters out bad groups/features.
}
\examples{

 
}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
