% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupFiltR.R
\name{GroupFiltR}
\alias{GroupFiltR}
\title{Group Filtering}
\usage{
GroupFiltR(XCMSobject, max_RT_width = NULL, max_ppm_diff = NULL,
  max_dupl_prop = NULL, max_missing_prop = NULL,
  thresholds_to_pass = NULL)
}
\arguments{
\item{XCMSobject}{An xcmsSet object after peak picking.}

\item{max_RT_width}{The maximal allowed RT width for a group}

\item{max_ppm_diff}{The maximal allowed ppm difference within a group}

\item{max_dupl_prop}{The maximal allowed proportion of duplicated peaks in the group}

\item{max_missing_prop}{The maximal allowed proportion of missing peaks in the group}

\item{thresholds_to_pass}{The number of thresholds to pass (default is all)}
}
\description{
After grouping and peak filling, this function filters out bad groups/features. 
There is the possibility of supplying groups of interest in which case the. 
TODO: param ToDo......ClassesOfInterest (optional) The classes to include in the calculation 
of group properties. 
param ToDo classVector (required if ClassesOfInterest provided) vector with classlabels.
param ToDo.......WithinGroupsOfInterest (default is FALSE), Option to choose to do the 
calculations within class instead of over the entire group/feature
}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
