% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meeseeks.RF.R
\name{Meeseeks.RF}
\alias{Meeseeks.RF}
\title{Easy Cross-Validated Random Forest}
\usage{
Meeseeks.RF(FeatureMatrix, GroupLabels, SampleLabels = NULL, nFolds = 10,
  nSims = 20, plot.out = TRUE, plot.type = "ROC", nCPU = -1,
  plotcol = NULL)
}
\arguments{
\item{FeatureMatrix}{The matrix of Features (obtained by using the xcms::groupval function). Matrix has to have columns for features and rows for samples.}

\item{GroupLabels}{The group labels. If not a factor a conversion will be applied.}

\item{SampleLabels}{(optional) unique sample identifier.}

\item{nFolds}{Number of cross validation folds.}

\item{nSims}{Number of simulations (every simulation has different folds)}

\item{plot.out}{Whether to print the ROC curve (default is TRUE).}

\item{plot.type}{Type of plot ourput. "ROC" for receiver operacter characteristic (default) or "PR" for precision-recall.}

\item{nCPU}{The number of cores to use (default is the maximum amount available minus 2)}

\item{plotcol}{(optional) colour to use for the plot}
}
\value{
A ROC plot (if plot.out = TRUE) and a list with 2 elements: 1) a data frame with the ROC plot data and 2) a matrix with the variable importance for each cross validated simulation (nFolds * nSims times).
}
\description{
This function quickly performs a cross-validated Random Forest classification on a data matrix.
}
\examples{

}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
