% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeseeks.SVM.R
\name{Meeseeks.SVM}
\alias{Meeseeks.SVM}
\title{Easy Cross-Validated SVM Binary Classification}
\usage{
Meeseeks.SVM(FeatureMatrix, GroupLabels, SampleLabels = NULL,
  nFolds = 10, nSims = 20, plot.out = TRUE, plot.type = "ROC",
  nCPU = -1, plotcol = NULL, svm.kernel = "linear",
  plottitle.extra = NULL, prediction_prob = TRUE, Nplotpoints = 501,
  ...)
}
\arguments{
\item{FeatureMatrix}{The matrix of Features (obtained by using the xcms::groupval function). Matrix has to have columns for features and rows for samples.}

\item{GroupLabels}{The group labels. If not a factor a conversion will be applied.}

\item{SampleLabels}{(optional) unique sample identifier.}

\item{nFolds}{Number of cross validation folds.}

\item{nSims}{Number of simulations (every simulation has different folds)}

\item{plot.out}{Whether to print the ROC curve (default is TRUE).}

\item{plot.type}{Type of plot output. "ROC" for receiver operator characteristic (default) or "PR" for precision-recall.}

\item{nCPU}{The number of cores to use (default is the maximum amount available minus 2)}

\item{plotcol}{(optional) colour to use for the plot}

\item{svm.kernel}{The kernel to be used for the svm (default is linear)}

\item{plottitle.extra}{Optional extra character string to be added to every plot title.}

\item{prediction_prob}{Whether to take the values for the TRUE prediction (default) or the FALSE prediction. Choosing FALSE inverts the ROC curve.}

\item{Nplotpoints}{The amount of points used to construct the plot.}

\item{...}{Extra paremeters to be passed along to \code{\link[e1071]{svm}}}
}
\value{
A ROC plot (if plot.out = TRUE) and a list with 2 elements: 1) a data frame with the ROC plot data and 2) a matrix with the variable importance for each cross validated simulation (nFolds * nSims times).
}
\description{
****** STILL CONTAINS BUGS (inverse ROC and PR) ****** This function quickly performs a cross-validated SVM classification on a data matrix.
}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
