% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.values.check.R
\name{missing.values.check}
\alias{missing.values.check}
\title{Missing value proportion checker}
\usage{
missing.values.check(DataMatrix, subclasses.list = NULL,
  prop.missing.threshold = 0.2, Na.or.numeric.limit = NA,
  labels = NULL, feature.orientation = "columns",
  groups.ok.threshold = 1)
}
\arguments{
\item{DataMatrix}{A matrix with feature groups as columns and samples for rows.}

\item{subclasses.list}{A list with either A) in each list entry a number of integers or 
numerics corresponding to the rows that form the subclasses to be checked, or B) in each 
list entry characters or factors corresponding to those found in the 'labels' attribute.}

\item{prop.missing.threshold}{The maximal proportion of missing values allowed in a single subclass.}

\item{Na.or.numeric.limit}{Define the missing values, this can be set to 'NA' or to a 
numerical value in which case all values below this value will be deemed missing.}

\item{labels}{A vector of labels which can be used to group the samples together in subclasses. 
This attribute is necessary if 'subclasses.list' consists of character or factor entries.}

\item{feature.orientation}{Indicates whether the features can be found in the 
columns (default) or in the rows. With the default setting every row corresponds 
to a sample, and every column to a feature/group}

\item{groups.ok.threshold}{The amount of subclasses (as defined in subclasses.list)}
}
\value{
Groups.that.passed A vector with the indices of the groups that passed the test 
for proportion of missing values.
}
\description{
function only lets groups of samples/subclasses pass which have a proportion of
missing values less than the allowed threshold
}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
