% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRSD.check.R
\name{mRSD.check}
\alias{mRSD.check}
\title{mRSD checker}
\usage{
mRSD.check(DataMatrix, subclasses.list = NULL, RSD.threshold = 0.4,
  ignore.missing.values = TRUE, Na.or.numeric.limit = NA, labels = NULL,
  feature.orientation = "columns", groups.ok.threshold = 1,
  include.RSD.matrix = FALSE)
}
\arguments{
\item{DataMatrix}{A matrix with feature groups as columns and samples for rows.}

\item{subclasses.list}{A list with either A) in each list entry a number of integers or 
numerics corresponding to the rows that form the subclasses to be checked, or B) in each 
list entry characters or factors corresponding to those found in the 'labels' attribute.}

\item{RSD.threshold}{The maximal RSD value (sd/mean) allowed in a single subclass.}

\item{ignore.missing.values}{Whether to ignore the missing values. If set to FALSE 
the calculation of the RSD will result in a lot of 'NAs' or 'NANs'. Only the features 
with no missing values will produce a numeric RSD.}

\item{Na.or.numeric.limit}{Define the missing values, this can be set to 'NA' or to a 
numerical value in which case all values below this value will be deemed missing.}

\item{labels}{A vector of labels which can be used to group the samples together in subclasses. 
This attribute is necessary if 'subclasses.list' consists of character or factor entries.}

\item{feature.orientation}{Indicates whether the features can be found in the columns (default) or 
in the rows. With the default setting every row corresponds to a sample, and every column to a feature/group}

\item{groups.ok.threshold}{The amount of subclasses (as defined in subclasses.list)}

\item{include.RSD.matrix}{If set to 'TRUE' the output will not only be a vector with the indices
of groups that passed the RSD check, but also the matrix with all RSD values is provided.}
}
\value{
RSD.results A vector with the indices of the groups that passed the test for RSD. 
If however the 'include.RSD.matrix' attribute is set to 'TRUE' the results will be 
a list with the vector of indices (list entry 1) and the matrix of RSD for each 
combination of subclass and feature (list entry 2)
}
\description{
function only lets groups of samples/subclasses pass which have a proportion of
missing values less than the allowed threshold
}
\examples{
# this is an example for 3 meaningless spectra
lengths_of_spectra <- c(100,150,120)
measurement_distance <- 0.01
starting_ppm_values <- c(8.7, 9.0, 9.0)
spectra <- list()
ppm_values <- list()
for (k in 1:3) {
    spectra[[k]] <- runif(lengths_of_spectra[k], min = 0, max = 10)
    
    # note the minus sign in the 'by' statement
    ppm_values[[k]] <- seq(from = starting_ppm_values[k], by = -measurement_distance, 
                           length.out = lengths_of_spectra[k])  
}
new.Data <- BuildRawDataMatrix(spectrum.list = spectra, ppm.list = ppm_values)
spectraMatrix <- new.Data$DataMatrix
ppmMatrix <- new.Data$ppmMatrix

}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
