% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.filter.R
\name{peak.filter}
\alias{peak.filter}
\title{Peak Filtering}
\usage{
peak.filter(XCMSobject, min.RT.width = NULL, max.RT.width = NULL,
  max.ppm.width = NULL, gauss.fail.threshold = NULL)
}
\arguments{
\item{XCMSobject}{An xcmsSet object after peak picking.}

\item{min.RT.width}{The minimal retention time width a peak must have}

\item{max.RT.width}{The maximal retention time width a peak may have}

\item{max.ppm.width}{The maximal allowed ppm width for a single peak}

\item{gauss.fail.threshold}{The threshold for the egauss XCMS parameter, below which the peak is rejected.}
}
\value{
filtered xcmsSet object
}
\description{
After peak picking this function filters out bad groups/features.
}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
