% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakFiltR.R
\name{PeakFiltR}
\alias{PeakFiltR}
\title{Peak Filtering}
\usage{
PeakFiltR(XCMSobject, min_RT_width = NULL, max_RT_width = NULL,
  max_ppm_width = NULL, gauss_fail_threshold = NULL)
}
\arguments{
\item{XCMSobject}{An xcmsSet object after peak picking.}

\item{min_RT_width}{The minimal retention time width a peak must have}

\item{max_RT_width}{The maximal retention time width a peak may have}

\item{max_ppm_width}{The maximal allowed ppm width for a single peak}

\item{gauss_fail_threshold}{The threshold for the egauss XCMS parameter, below which the peak is rejected.}
}
\value{
filtered xcmsSet object
}
\description{
After peak picking this function filters out bad groups/features.
}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
