% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.plots.peaks.R
\name{QC.plots.peaks}
\alias{QC.plots.peaks}
\title{QC plots for peak data}
\usage{
QC.plots.peaks(XCMSobject, include.postfill.plots = FALSE)
}
\arguments{
\item{XCMSobject}{An xcmsSet object}

\item{include.postfill.plots}{Whether to include the peaks added after filling (only if available)}
}
\value{
a QC plot for nr of peaks
}
\description{
Mean number of peaks are plotted for the peak data
}
\examples{


}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
