% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plots_features.R
\name{QC_plots_features}
\alias{QC_plots_features}
\title{QC plots for feature data}
\usage{
QC_plots_features(FeatureMatrix, classVector, COI = NULL,
  NA_numeric_limit = NULL, plottitle = NULL, BOI_vector = NULL,
  BOI = NULL, RDSplot_text_adjustheigt = 0)
}
\arguments{
\item{FeatureMatrix}{the matrix of Features (obtained by using the xcms::groupval function). Matrix has to have columns for features and rows for samples.}

\item{classVector}{The vector of sample classes.}

\item{COI}{(optional) In case only a specific class is of interest.}

\item{NA_numeric_limit}{If missing values are not indicated as NA but with a numeric value like 0, supply the numerical value under which values are considered missing.}

\item{plottitle}{(optional) Plot title}

\item{BOI_vector}{(optional) In case the RSD's need only be calculated for a specific subset of batches, provide the character/factor vector witch batches here (same length as the number of rows in FeatureMatrix).}

\item{BOI}{(optional) The name of the batch of interest. BOI_vector has to be supplied as well}

\item{RDSplot_text_adjustheigt}{Height adjustment for the count number in the RSD boxplot visuzlization.}
}
\value{
several QC plots
}
\description{
RSD's for each class are plotted.
}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
