% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plots_peaks.R
\name{QC_plots_peaks}
\alias{QC_plots_peaks}
\title{QC plots for peak data}
\usage{
QC_plots_peaks(XCMSobject, include_postfill_plots = FALSE,
  className = NULL, plottitle = NULL)
}
\arguments{
\item{XCMSobject}{An xcmsSet object}

\item{className}{In case there are multiple class definitions in the XCMSobject@phenoData object, the one of interest can be specified here. If not the code will ask.}

\item{plottitle}{(optional) The title to place above the plot(s).}

\item{include.postfill.plots}{Whether to include the peaks added after filling (only if available)}
}
\value{
A QC plot for nr of peaks
}
\description{
Mean number of peaks are plotted for the peak data
}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
