% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTcorrChecker.R
\name{RTcorrChecker}
\alias{RTcorrChecker}
\title{NOT FINISHED!!!!!! Function to compare different RT corrections by looking at the most shifted peaks from in method}
\usage{
RTcorrChecker(RTcorr1.xcmsGrouped = test.rtcorrs.grouped[[1]],
  RTcorr2.xcmsGrouped = test.rtcorrs.obiwarp.grouped[[1]],
  plotTitle.method1 = "Loess.correction, span = 1",
  plotTitle.method2 = "Obiwarp correction, respons = 10",
  RTdiff.cutoff.value = NULL, RTdiff.cutoff.quantile = NULL)
}
\arguments{
\item{RTcorr1.xcmsGrouped}{XCMS object of first RT correction method (post RT correction and subsequent grouping)}

\item{RTcorr2.xcmsGrouped}{XCMS object of second RT correction method (post RT correction and subsequent grouping)}
}
\value{

}
\description{
Filter out isotopes based on the PUTMEDID concept
}
\examples{


}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
