% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Std.matcher.R
\name{Std.matcher}
\alias{Std.matcher}
\title{Standards Locater}
\usage{
Std.matcher(Standards.mz, Standards.RT, Data, Standards.ratios = NULL,
  max.ppm = NULL, max.mz = NULL, max.RT.narrow = NULL,
  max.RT.wide = NULL, accept.ppm.diff = NULL)
}
\arguments{
\item{Standards.mz}{A vector (single standards) or matrix (isotopic profiles) with the standards.}

\item{Standards.RT}{A vector with the RT values}

\item{Data}{This can be either a) a matrix with 2 or 3* columns (mz, RT, Intensity*) without missing values, b) an xcmsSet object post grouping (the groups are features, intensity ratio's can be checked) or c) an XCMS@groups matrix object (if only this is supplied there will be no intensity ratio check). Note that an XCMS@peaks object is also supported but the function only returns 1 match per standard (the closest one, or the first of a number of equally close ones)}

\item{Standards.ratios}{Optional matrix. Matrix with the intensity ratio's}

\item{max.ppm}{maximal ppm distance}

\item{max.mz}{maximal mz distance (not necessary if ppm supplied)}

\item{max.RT.narrow}{prefered maximal RT distance}

\item{max.RT.wide}{extended RT search area. A peak is only accepted here when the improvement in ppm difference is at least accept.ppm.diff}

\item{accept.ppm.diff}{The improvement in ppm difference a peak in the wider RT area has to have for it to be accepted}
}
\value{
A list with 3 vector/matrix entries (the same size as the supplied vector/matrix with standards). 
The first are the matched peaks/groups. The second the ppm differences. The third the RT differences.
}
\description{
This function searces for XCMS objects (peaks, groups) that match with the provided standards (optionally with isotopic profiles).
}
\examples{


}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
