% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unStructureBatches.R
\name{unStructureBatches}
\alias{unStructureBatches}
\title{obtain all datasets in a single list for easy processing with purrr}
\usage{
unStructureBatches(xcmsSets.Post.batchSplit, BatchProcessingStructure)
}
\arguments{
\item{xcmsSets.Post.batchSplit}{list of xcmsSet objects after Batch splitting (BatchSplitR function)}

\item{BatchProcessingStructure}{The batch processing structure (obtained with getBatchStructure)}
}
\value{
unStructuredBatchList a list with every individual batch in a list element
}
\description{
obtain all datasets in a single list for easy processing with purrr
}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
