% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamss.parallel.R
\name{yamms.parallel}
\alias{yamms.parallel}
\title{parallel beakedpi peak picker}
\usage{
yamms.parallel(Datafiles, yamms.colData, nCPU = -1, mzRANGE,
  slicewidth = 100, sliceoverlap = 5, dbandwidth = c(0.01, 10),
  dgridstep = c(0.01, 1), outfileDens = NULL, dortalign = FALSE)
}
\arguments{
\item{DataMatrix}{A matrix with feature groups as columns and samples for rows.}

\item{subclasses.list}{A list with either A) in each list entry a number of integers or 
numerics corresponding to the rows that form the subclasses to be checked, or B) in each 
list entry characters or factors corresponding to those found in the 'labels' attribute.}
}
\value{
yammsList - alist with the data matrix, groupinfo and sample classes
}
\description{
function to parallelize the bakedpi algorithm
}
\examples{

}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
