% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDAwrapper.R
\name{PLSDAwrapper}
\alias{PLSDAwrapper}
\title{Function to split up xcms object by catch to perform a within batch analysis and merge the results at the end}
\usage{
PLSDAwrapper(plsdaMatrix, classLabels, plotMarks = NULL,
  plotOutputs = FALSE)
}
\arguments{
\item{plsdaMatrix}{The matrix to be used in the PLSDA}

\item{classLabels}{The class labels. The length must correspond to the amount of rows in the plsdaMatrix}

\item{plotMarks}{(optional) A character vector with plot labels (same length as classLabels).}

\item{plotOutputs}{(logical) Whether to plot the PCA scores and loadings.}
}
\value{
varimportance.PLSDA The vector with variable importances
}
\description{
Function to split up xcms object by catch to perform a within batch analysis and merge the results at the end
}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
