% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.plots.features.R
\name{QC.plots.features}
\alias{QC.plots.features}
\title{QC plots for feature data}
\usage{
QC.plots.features(FeatureMatrix, XCMSobject)
}
\arguments{
\item{FeatureMatrix}{the matrix of Features (obtained by using the xcms::groupval function). Matrix has to have columns for features and rows for samples.}

\item{XCMSobject}{An xcmsSet object}
}
\value{
several QC plots
}
\description{
Mean number of peaks are plotted for the peak data
}
\examples{


}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
