% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.plots.features.R
\name{QC.plots.features}
\alias{QC.plots.features}
\title{QC plots for feature data}
\usage{
QC.plots.features(FeatureMatrix, classVector, COI = NULL,
  NA.numeric.limit = NULL, plottitle = NULL, BOI.vector = NULL,
  BOI = NULL)
}
\arguments{
\item{FeatureMatrix}{the matrix of Features (obtained by using the xcms::groupval function). Matrix has to have columns for features and rows for samples.}

\item{classVector}{The vector of sample classes.}

\item{COI}{(optional) In case only a specific class is of interest.}

\item{NA.numeric.limit}{If missing values are not indicated as NA but with a numeric value like 0, supply the numerical value under which values are considered missing.}

\item{plottitle}{(optional) Plot title}

\item{BOI.vector}{(optional) In case the RSD's need only be calculated for a specific subset of batches, provide the character/factor vector witch batches here (same length as the number of rows in FeatureMatrix).}

\item{BOI}{(optional) The name of the batch of interest. BOI.vector has to be supplied as well}
}
\value{
several QC plots
}
\description{
Mean number of peaks are plotted for the peak data
}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
